Share some implementation
