+++ src/lib.rs
@@ -388,11 +388,11 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn rust() -> Result<()> {
+    async fn rust_wit_bindgen_02() -> Result<()> {
         run_spin(
             &fs::read(concat!(
                 env!("OUT_DIR"),
-                "/wasm32-wasi/release/rust_case.wasm"
+                "/wasm32-wasi/release/rust_case_02.wasm"
             ))
             .await?,
         )
@@ -400,7 +400,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn rust_07() -> Result<()> {
+    async fn rust_wit_bindgen_07() -> Result<()> {
         run_spin(
             &fs::read(concat!(
                 env!("OUT_DIR"),
