+++ tests/case-helper/Cargo.toml
@@ -6,3 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = "1.0.44"
+clap = { version = "3.2.20", features = ["derive", "env"] }
+getrandom = "0.2.7"
