+++ tests/rust-case-0.2/src/lib.rs
@@ -1,13 +1,11 @@
-use anyhow::{anyhow, bail, Result};
-use clap::{Parser, Subcommand};
+use anyhow::{bail, Result};
+use case_helper::Command;
 use outbound_redis::RedisParameter;
 use spin_http::{Method, Request, Response};
 use std::{
-    env, error, fmt,
-    fs::{self, File},
+    error, fmt,
     io::{self, Write},
-    iter, str,
-    time::SystemTime,
+    str,
 };
 
 wit_bindgen_rust::export!("../wit/spin-http.wit");
@@ -154,183 +152,6 @@ impl fmt::Display for key_value::Error {
 
 impl error::Error for key_value::Error {}
 
-fn parse_pg(param: &str) -> Result<outbound_pg::ParameterValue> {
-    use outbound_pg::ParameterValue as PV;
-
-    Ok(if param == "null" {
-        PV::DbNull
-    } else {
-        let (type_, value) = param
-            .split_once(':')
-            .ok_or_else(|| anyhow!("expected ':' in {param}"))?;
-
-        match type_ {
-            "boolean" => PV::Boolean(value.parse()?),
-            "int8" => PV::Int8(value.parse()?),
-            "int16" => PV::Int16(value.parse()?),
-            "int32" => PV::Int32(value.parse()?),
-            "int64" => PV::Int64(value.parse()?),
-            "uint8" => PV::Uint8(value.parse()?),
-            "uint16" => PV::Uint16(value.parse()?),
-            "uint32" => PV::Uint32(value.parse()?),
-            "uint64" => PV::Uint64(value.parse()?),
-            "floating32" => PV::Floating32(value.parse()?),
-            "floating64" => PV::Floating64(value.parse()?),
-            "str" => PV::Str(value),
-            "binary" => PV::Binary(value.as_bytes()),
-            _ => bail!("unknown parameter type: {type_}"),
-        }
-    })
-}
-
-fn parse_mysql(param: &str) -> Result<outbound_mysql::ParameterValue> {
-    use outbound_mysql::ParameterValue as PV;
-
-    Ok(if param == "null" {
-        PV::DbNull
-    } else {
-        let (type_, value) = param
-            .split_once(':')
-            .ok_or_else(|| anyhow!("expected ':' in {param}"))?;
-
-        match type_ {
-            "boolean" => PV::Boolean(value.parse()?),
-            "int8" => PV::Int8(value.parse()?),
-            "int16" => PV::Int16(value.parse()?),
-            "int32" => PV::Int32(value.parse()?),
-            "int64" => PV::Int64(value.parse()?),
-            "uint8" => PV::Uint8(value.parse()?),
-            "uint16" => PV::Uint16(value.parse()?),
-            "uint32" => PV::Uint32(value.parse()?),
-            "uint64" => PV::Uint64(value.parse()?),
-            "floating32" => PV::Floating32(value.parse()?),
-            "floating64" => PV::Floating64(value.parse()?),
-            "str" => PV::Str(value),
-            "binary" => PV::Binary(value.as_bytes()),
-            _ => bail!("unknown parameter type: {type_}"),
-        }
-    })
-}
-
-#[derive(Debug, Parser)]
-#[clap(author, version, about)]
-pub struct Cli {
-    #[clap(subcommand)]
-    command: Command,
-}
-
-#[derive(Debug, Subcommand)]
-enum Command {
-    Config {
-        key: String,
-    },
-    Http {
-        url: String,
-    },
-    RedisPublish {
-        address: String,
-        key: String,
-        value: String,
-    },
-    RedisSet {
-        address: String,
-        key: String,
-        value: String,
-    },
-    RedisGet {
-        address: String,
-        key: String,
-    },
-    RedisIncr {
-        address: String,
-        key: String,
-    },
-    RedisDel {
-        address: String,
-        keys: Vec<String>,
-    },
-    RedisSadd {
-        address: String,
-        key: String,
-        params: Vec<String>,
-    },
-    RedisSrem {
-        address: String,
-        key: String,
-        params: Vec<String>,
-    },
-    RedisSmembers {
-        address: String,
-        key: String,
-    },
-    RedisExecute {
-        address: String,
-        command: String,
-        params: Vec<String>,
-    },
-    PostgresExecute {
-        address: String,
-        statement: String,
-        params: Vec<String>,
-    },
-    PostgresQuery {
-        address: String,
-        statement: String,
-        params: Vec<String>,
-    },
-    MysqlExecute {
-        address: String,
-        statement: String,
-        params: Vec<String>,
-    },
-    MysqlQuery {
-        address: String,
-        statement: String,
-        params: Vec<String>,
-    },
-    KeyValueOpen {
-        name: String,
-    },
-    KeyValueGet {
-        store: u32,
-        key: String,
-    },
-    KeyValueSet {
-        store: u32,
-        key: String,
-        value: String,
-    },
-    KeyValueDelete {
-        store: u32,
-        key: String,
-    },
-    KeyValueExists {
-        store: u32,
-        key: String,
-    },
-    KeyValueGetKeys {
-        store: u32,
-    },
-    KeyValueClose {
-        store: u32,
-    },
-    WasiEnv {
-        key: String,
-    },
-    WasiEpoch,
-    WasiRandom,
-    WasiStdio,
-    WasiRead {
-        file_name: String,
-    },
-    WasiReaddir {
-        dir_name: String,
-    },
-    WasiStat {
-        file_name: String,
-    },
-}
-
 fn dispatch(body: Option<Vec<u8>>) -> Response {
     match execute(body) {
         Ok(()) => {
@@ -353,12 +174,7 @@ fn dispatch(body: Option<Vec<u8>>) -> Response {
 }
 
 fn execute(body: Option<Vec<u8>>) -> Result<()> {
-    let body = body.ok_or_else(|| anyhow!("empty request body"))?;
-    let command = iter::once("<wasm module>")
-        .chain(str::from_utf8(&body)?.split("%20"))
-        .collect::<Vec<_>>();
-
-    match &Cli::try_parse_from(command)?.command {
+    match &Command::extract(body)? {
         Command::Config { key } => {
             spin_config::get_config(key)?;
         }
@@ -537,59 +353,68 @@ fn execute(body: Option<Vec<u8>>) -> Result<()> {
             key_value::close(*store);
         }
 
-        Command::WasiEnv { key } => print!("{}", env::var(key)?),
-
-        Command::WasiEpoch => print!(
-            "{}",
-            SystemTime::now()
-                .duration_since(SystemTime::UNIX_EPOCH)?
-                .as_millis()
-        ),
-
-        Command::WasiRandom => {
-            let mut buffer = [0u8; 8];
-            getrandom::getrandom(&mut buffer).map_err(|_| anyhow!("getrandom error"))?;
-        }
-
-        Command::WasiStdio => {
-            io::copy(&mut io::stdin().lock(), &mut io::stdout().lock())?;
-        }
-
-        Command::WasiRead { file_name } => {
-            io::copy(&mut File::open(file_name)?, &mut io::stdout().lock())?;
-        }
-
-        Command::WasiReaddir { dir_name } => {
-            let mut comma = false;
-            for entry in fs::read_dir(dir_name)? {
-                if comma {
-                    print!(",");
-                } else {
-                    comma = true;
-                }
-
-                print!(
-                    "{}",
-                    entry?
-                        .file_name()
-                        .to_str()
-                        .ok_or_else(|| anyhow!("non-UTF-8 file name"))?
-                );
-            }
-        }
-
-        Command::WasiStat { file_name } => {
-            let metadata = fs::metadata(file_name)?;
-            print!(
-                "length:{},modified:{}",
-                metadata.len(),
-                metadata
-                    .modified()?
-                    .duration_since(SystemTime::UNIX_EPOCH)?
-                    .as_millis()
-            );
-        }
+        Command::WasiEnv { key } => Command::env(key.clone())?,
+        Command::WasiEpoch => Command::epoch()?,
+        Command::WasiRandom => Command::random()?,
+        Command::WasiStdio => Command::stdio()?,
+        Command::WasiRead { file_name } => Command::read(file_name.clone())?,
+        Command::WasiReaddir { dir_name } => Command::read_dir(dir_name.clone())?,
+        Command::WasiStat { file_name } => Command::stat(file_name.clone())?,
     }
 
     Ok(())
+}
+
+fn parse_pg(param: &str) -> Result<outbound_pg::ParameterValue> {
+    use outbound_pg::ParameterValue as PV;
+
+    Ok(if param == "null" {
+        PV::DbNull
+    } else {
+        let (type_, value) = case_helper::split_param(param)?;
+
+        match type_ {
+            "boolean" => PV::Boolean(value.parse()?),
+            "int8" => PV::Int8(value.parse()?),
+            "int16" => PV::Int16(value.parse()?),
+            "int32" => PV::Int32(value.parse()?),
+            "int64" => PV::Int64(value.parse()?),
+            "uint8" => PV::Uint8(value.parse()?),
+            "uint16" => PV::Uint16(value.parse()?),
+            "uint32" => PV::Uint32(value.parse()?),
+            "uint64" => PV::Uint64(value.parse()?),
+            "floating32" => PV::Floating32(value.parse()?),
+            "floating64" => PV::Floating64(value.parse()?),
+            "str" => PV::Str(value),
+            "binary" => PV::Binary(value.as_bytes()),
+            _ => bail!("unknown parameter type: {type_}"),
+        }
+    })
+}
+
+fn parse_mysql(param: &str) -> Result<outbound_mysql::ParameterValue> {
+    use outbound_mysql::ParameterValue as PV;
+
+    Ok(if param == "null" {
+        PV::DbNull
+    } else {
+        let (type_, value) = case_helper::split_param(param)?;
+
+        match type_ {
+            "boolean" => PV::Boolean(value.parse()?),
+            "int8" => PV::Int8(value.parse()?),
+            "int16" => PV::Int16(value.parse()?),
+            "int32" => PV::Int32(value.parse()?),
+            "int64" => PV::Int64(value.parse()?),
+            "uint8" => PV::Uint8(value.parse()?),
+            "uint16" => PV::Uint16(value.parse()?),
+            "uint32" => PV::Uint32(value.parse()?),
+            "uint64" => PV::Uint64(value.parse()?),
+            "floating32" => PV::Floating32(value.parse()?),
+            "floating64" => PV::Floating64(value.parse()?),
+            "str" => PV::Str(value),
+            "binary" => PV::Binary(value.as_bytes()),
+            _ => bail!("unknown parameter type: {type_}"),
+        }
+    })
 }
